+++ lib/controllers/controllers.dart
@@ -4,3 +4,4 @@
 export 'zen_controller.dart' show ZenController;
 export 'zen_route_observer.dart';
 export 'zen_controller_scope.dart';
+export 'zen_service.dart';